refactor: 프로젝트 재구성
